// lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  TENANT: 'TENANT',
  OWNER: 'OWNER',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const UserStatus = {
  OPEN: 'OPEN',
  NOT_VERIFIED: 'NOT_VERIFIED',
  VERIFIED: 'VERIFIED',
  SUSPENDED: 'SUSPENDED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const Gender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE',
  OTHER: 'OTHER'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const Religion = {
  ISLAM: 'ISLAM',
  HINDUISM: 'HINDUISM',
  CHRISTIANITY: 'CHRISTIANITY',
  BUDDHISM: 'BUDDHISM',
  OTHER: 'OTHER'
} as const

export type Religion = (typeof Religion)[keyof typeof Religion]


export const HouseholdType = {
  BACHELOR: 'BACHELOR',
  FAMILY: 'FAMILY'
} as const

export type HouseholdType = (typeof HouseholdType)[keyof typeof HouseholdType]


export const PropertyType = {
  APARTMENT: 'APARTMENT',
  HOUSE: 'HOUSE',
  ROOM: 'ROOM'
} as const

export type PropertyType = (typeof PropertyType)[keyof typeof PropertyType]


export const PetPolicy = {
  ALLOWED: 'ALLOWED',
  NOT_ALLOWED: 'NOT_ALLOWED'
} as const

export type PetPolicy = (typeof PetPolicy)[keyof typeof PetPolicy]


export const PropertyStatus = {
  PENDING: 'PENDING',
  AVAILABLE: 'AVAILABLE',
  RENTED: 'RENTED',
  REJECTED: 'REJECTED'
} as const

export type PropertyStatus = (typeof PropertyStatus)[keyof typeof PropertyStatus]


export const BookingStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus]